requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0 